cdppro/restservices/api/cdpweb/requirements.txt
djangorestframework==3.9.2
django-rest-swagger==2.2.0
django-cors-headers==3.0.2
mysql-connector-python==8.0.15
mysqlclient==1.4.2
html5lib==1.0.1
pandas==0.24.2
#uwsgi==2.0.18
applicationinsights==0.11.9
